requirements.txt
Flask==3.0.0
requests==2.31.0
opencv-python==4.8.1.78
pyserial==3.5
fonttools>=4.43.0
hershey-fonts>=0.0.2